+++ CHANGES.rst
@@ -30,6 +30,7 @@ Release history
 - Provide a more informative error message if Layer ``shape_in``/``shape_out`` contains
   undefined (``None``) elements. (`#119`_)
 - Fixed bug in ``Converter`` when source model contains duplicate nodes. (`#119`_)
+- Fixed bug in ``Converter`` for ``Concatenate`` layers with ``axis != 1``. (`#119`_)
 
 .. _#119: https://github.com/nengo/nengo-dl/pull/119